move utility to gencommon
